change the type of Genre to decimal

+++ Back-End/MoviesServer/Models/Movie.cs
@@ -10,13 +10,13 @@ namespace MoviesServer.Models
         public string Director { get; set; }
         public int ReleaseYear { get; set; }
         public string Genre { get; set; }
-        public double Rating { get; set; }
+        public decimal Rating { get; set; }
         public int Runtime { get; set; }
         public string Description { get; set; }
         public List<string> Responses { get; set; }
         public string Image { get; set; }
 
-        public Movie(int movieID, string title, string director, int releaseYear, string genre, double rating, int runtime, string description, List<string> responses, string image)
+        public Movie(int movieID, string title, string director, int releaseYear, string genre, decimal rating, int runtime, string description, List<string> responses, string image)
         {
             MovieID = movieID;
             Title = title;
